Fix showing uninstaller in "remove programs"
The uninstaller wasn't showing up in Add/Remove Programs because we were
trying to put the registry keys in HKEY_LOCAL_MACHINE, but the installer
isn't running as administrator so it had no permissions there.

Put the registry keys in HKEY_CURRENT_USER instead, which makes more sense
since we're installing in a user-local directory.